chore: move SmartDocuments rest to ZaakafhandelParametersRESTService
